refactor : make 함수 파일로나누기
